# docs/openapi.yaml
openapi: 3.0.1
info:
  title: BLPS - Job Vacancy Management API
  description: |-
    REST API for job vacancy management system (hh.ru-like).

    ## Features
    - User registration and JWT authentication
    - Vacancy management (CRUD operations)
    - Tariff selection and vacancy publishing (BPMN workflow)
    - Role-based access control (EMPLOYER, ADMIN)

    ## Workflow (BPMN)
    1. Create vacancy (DRAFT status)
    2. Select tariff
    3. Publish vacancy (PUBLISHED status)
    4. Archive vacancy when needed (ARCHIVED status)
  contact:
    name: BLPS Project
    email: support@blps.example.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
tags:
- name: User & Admin
  description: User profile and admin endpoints
- name: Authentication
  description: Authentication and registration endpoints
- name: Vacancies
  description: Vacancy management endpoints
- name: Tariffs
  description: Tariff management endpoints
paths:
  /api/v1/vacancies/{id}:
    get:
      tags:
      - Vacancies
      summary: Get vacancy by ID
      description: Get vacancy details by ID (public)
      operationId: getVacancyById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VacancyResponse"
    put:
      tags:
      - Vacancies
      summary: Update vacancy
      description: Update existing vacancy (owner or admin)
      operationId: updateVacancy
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVacancyRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VacancyResponse"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Vacancies
      summary: Delete vacancy
      description: Delete vacancy (owner or admin)
      operationId: deleteVacancy
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
      security:
      - bearerAuth: []
  /api/v1/tariffs/{id}:
    get:
      tags:
      - Tariffs
      summary: Get tariff by ID
      description: Get tariff details by ID (public)
      operationId: getTariffById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TariffResponse"
    put:
      tags:
      - Tariffs
      summary: Update tariff
      description: Update existing tariff (admin only)
      operationId: updateTariff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTariffRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TariffResponse"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Tariffs
      summary: Delete tariff
      description: Delete tariff by ID (admin only)
      operationId: deleteTariff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
      security:
      - bearerAuth: []
  /api/v1/vacancies:
    get:
      tags:
      - Vacancies
      summary: Get all vacancies
      description: Get paginated list of all vacancies (public)
      operationId: getAllVacancies
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - DRAFT
          - PUBLISHED
          - ARCHIVED
          - CLOSED
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageVacancyResponse"
    post:
      tags:
      - Vacancies
      summary: Create vacancy
      description: Create a new vacancy draft
      operationId: createVacancy
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVacancyRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VacancyResponse"
      security:
      - bearerAuth: []
  /api/v1/tariffs:
    get:
      tags:
      - Tariffs
      summary: Get all tariffs
      description: Get paginated list of all tariffs (public)
      operationId: getAllTariffs
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageTariffResponse"
    post:
      tags:
      - Tariffs
      summary: Create tariff
      description: Create a new tariff (admin only)
      operationId: createTariff
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTariffRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TariffResponse"
      security:
      - bearerAuth: []
  /api/v1/auth/register:
    post:
      tags:
      - Authentication
      summary: Register new employer
      description: Register a new employer account
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/v1/auth/refresh:
    post:
      tags:
      - Authentication
      summary: Refresh token
      description: Refresh access token using refresh token
      operationId: refreshToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/v1/auth/login:
    post:
      tags:
      - Authentication
      summary: Login
      description: Login with email and password
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/v1/vacancies/{id}/tariff:
    patch:
      tags:
      - Vacancies
      summary: Select tariff
      description: "Select tariff for vacancy (BPMN: Draft -> Payment)"
      operationId: selectTariff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: tariffId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VacancyResponse"
      security:
      - bearerAuth: []
  /api/v1/vacancies/{id}/publish:
    patch:
      tags:
      - Vacancies
      summary: Publish vacancy
      description: "Publish vacancy (BPMN: Draft -> Published)"
      operationId: publishVacancy
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VacancyResponse"
      security:
      - bearerAuth: []
  /api/v1/vacancies/{id}/archive:
    patch:
      tags:
      - Vacancies
      summary: Archive vacancy
      description: "Archive vacancy (BPMN: Published -> Archived)"
      operationId: archiveVacancy
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VacancyResponse"
      security:
      - bearerAuth: []
  /api/v1/vacancies/my:
    get:
      tags:
      - Vacancies
      summary: Get my vacancies
      description: Get paginated list of current user's vacancies
      operationId: getMyVacancies
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - DRAFT
          - PUBLISHED
          - ARCHIVED
          - CLOSED
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageVacancyResponse"
      security:
      - bearerAuth: []
  /api/v1/users/me:
    get:
      tags:
      - User & Admin
      summary: Get current user
      description: Get current authenticated user profile
      operationId: getCurrentUser
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
      security:
      - bearerAuth: []
  /api/v1/admin/users/{userId}:
    delete:
      tags:
      - User & Admin
      summary: Delete user
      description: "Delete user and all their vacancies (admin only, cascaded deletion)"
      operationId: deleteUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
      security:
      - bearerAuth: []
components:
  schemas:
    UpdateVacancyRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        experienceLevel:
          type: string
          enum:
          - NO_EXPERIENCE
          - ONE_TO_THREE
          - THREE_TO_SIX
          - MORE_THAN_SIX
        salaryFrom:
          type: number
        salaryTo:
          type: number
        employmentType:
          type: string
          enum:
          - FULL_TIME
          - PART_TIME
          - PROJECT
          - INTERNSHIP
          - VOLUNTEER
        workFormat:
          type: string
          enum:
          - REMOTE
          - OFFICE
          - HYBRID
        employmentFormat:
          type: string
          enum:
          - EMPLOYMENT_CONTRACT
          - CIVIL_CONTRACT
          - SELF_EMPLOYED
        workSchedule:
          type: string
          enum:
          - NINE_TO_EIGHTEEN
          - TEN_TO_NINETEEN
          - EIGHT_TO_SEVENTEEN
          - FLEXIBLE
          - SHIFT
          - REMOTE_FLEXIBLE
        city:
          type: string
        address:
          type: string
        companyDescription:
          type: string
        additionalSkills:
          type: array
          items:
            type: string
    VacancyResponse:
      required:
      - additionalSkills
      - city
      - createdAt
      - description
      - employerId
      - employmentFormat
      - employmentType
      - experienceLevel
      - id
      - status
      - title
      - updatedAt
      - workFormat
      - workSchedule
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        experienceLevel:
          type: string
          enum:
          - NO_EXPERIENCE
          - ONE_TO_THREE
          - THREE_TO_SIX
          - MORE_THAN_SIX
        salaryFrom:
          type: number
        salaryTo:
          type: number
        employmentType:
          type: string
          enum:
          - FULL_TIME
          - PART_TIME
          - PROJECT
          - INTERNSHIP
          - VOLUNTEER
        workFormat:
          type: string
          enum:
          - REMOTE
          - OFFICE
          - HYBRID
        employmentFormat:
          type: string
          enum:
          - EMPLOYMENT_CONTRACT
          - CIVIL_CONTRACT
          - SELF_EMPLOYED
        workSchedule:
          type: string
          enum:
          - NINE_TO_EIGHTEEN
          - TEN_TO_NINETEEN
          - EIGHT_TO_SEVENTEEN
          - FLEXIBLE
          - SHIFT
          - REMOTE_FLEXIBLE
        city:
          type: string
        address:
          type: string
        companyDescription:
          type: string
        additionalSkills:
          type: array
          items:
            type: string
        status:
          type: string
          enum:
          - DRAFT
          - PUBLISHED
          - ARCHIVED
          - CLOSED
        employerId:
          type: string
        tariffId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
    UpdateTariffRequest:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        durationDays:
          type: integer
          format: int32
        description:
          type: string
    TariffResponse:
      required:
      - description
      - durationDays
      - id
      - name
      - price
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
        durationDays:
          type: integer
          format: int32
        description:
          type: string
    CreateVacancyRequest:
      required:
      - additionalSkills
      - city
      - description
      - employmentFormat
      - employmentType
      - experienceLevel
      - title
      - workFormat
      - workSchedule
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        experienceLevel:
          type: string
          enum:
          - NO_EXPERIENCE
          - ONE_TO_THREE
          - THREE_TO_SIX
          - MORE_THAN_SIX
        salaryFrom:
          type: number
        salaryTo:
          type: number
        employmentType:
          type: string
          enum:
          - FULL_TIME
          - PART_TIME
          - PROJECT
          - INTERNSHIP
          - VOLUNTEER
        workFormat:
          type: string
          enum:
          - REMOTE
          - OFFICE
          - HYBRID
        employmentFormat:
          type: string
          enum:
          - EMPLOYMENT_CONTRACT
          - CIVIL_CONTRACT
          - SELF_EMPLOYED
        workSchedule:
          type: string
          enum:
          - NINE_TO_EIGHTEEN
          - TEN_TO_NINETEEN
          - EIGHT_TO_SEVENTEEN
          - FLEXIBLE
          - SHIFT
          - REMOTE_FLEXIBLE
        city:
          type: string
        address:
          type: string
        companyDescription:
          type: string
        additionalSkills:
          type: array
          items:
            type: string
    CreateTariffRequest:
      required:
      - description
      - durationDays
      - name
      - price
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        durationDays:
          type: integer
          format: int32
        description:
          type: string
    RegisterRequest:
      required:
      - companyName
      - email
      - password
      type: object
      properties:
        email:
          type: string
        password:
          maxLength: 2147483647
          minLength: 6
          type: string
        companyName:
          type: string
    AuthResponse:
      required:
      - accessToken
      - email
      - refreshToken
      - role
      - userId
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        userId:
          type: string
        email:
          type: string
        role:
          type: string
    RefreshTokenRequest:
      required:
      - refreshToken
      type: object
      properties:
        refreshToken:
          type: string
    LoginRequest:
      required:
      - email
      - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageVacancyResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/VacancyResponse"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
    UserResponse:
      required:
      - companyName
      - email
      - id
      - role
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        companyName:
          type: string
        role:
          type: string
    PageTariffResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/TariffResponse"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        empty:
          type: boolean
  securitySchemes:
    bearerAuth:
      type: http
      description: Enter JWT token obtained from /api/v1/auth/login or /api/v1/auth/register
      scheme: bearer
      bearerFormat: JWT